% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build an Object Iteratively with a Single List}
\usage{
build(obj, .x, .f, ...)
}
\arguments{
\item{obj}{The initial object to be modified.}

\item{.x}{A list of elements to be passed as the second argument to \code{.f}.}

\item{.f}{The function to apply to \code{obj} and elements of \code{.x}.}

\item{...}{Additional arguments to pass to \code{.f}.}
}
\value{
The modified object.
}
\description{
This function takes an initial object and iteratively modifies it using a function and a list of arguments.
}
\examples{
\dontrun{
build(0, 1:5, `+`)
}
}
