% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build2}
\alias{build2}
\title{Build an Object Iteratively with Two Lists}
\usage{
build2(obj, .x, .y, .f, ...)
}
\arguments{
\item{obj}{The initial object to be modified.}

\item{.x}{A list of elements to be passed as the second argument to \code{.f}.}

\item{.y}{A list of elements to be passed as the third argument to \code{.f}.}

\item{.f}{The function to apply to \code{obj}, elements of \code{.x}, and elements of \code{.y}.}

\item{...}{Additional arguments to pass to \code{.f}.}
}
\value{
The modified object.
}
\description{
This function takes an initial object and iteratively modifies it using a function and two lists of arguments.
}
\examples{
\dontrun{
build2(0, 1:3, 4:6, `+`)
}
}
