% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{pbuild}
\alias{pbuild}
\title{Build an Object Iteratively with a List of Lists}
\usage{
pbuild(obj, .l, .f, ...)
}
\arguments{
\item{obj}{The initial object to be modified.}

\item{.l}{A list of lists, where each inner list contains elements to be passed as arguments to \code{.f}.}

\item{.f}{The function to apply to \code{obj} and elements of \code{.l}.}

\item{...}{Additional arguments to pass to \code{.f}.}
}
\value{
The modified object.
}
\description{
This function takes an initial object and iteratively modifies it using a function and a list of lists of arguments.
}
\examples{
\dontrun{
pbuild(0, list(1:3, 4:6), `+`)
}
}
